add and use charts colors
